requirements.txt
opencv-python
torch
torchvision
scikit-learn
numpy
pandas
streamlit
Pillow
flask
flask-cors
